src: Refactor user space PM commands.
Refactor the multipath-tcp.org kernel path management commands
implementations so that they may be leveraged by the corresponding and
pending upstream kernel user space commands.
